// job-site/src/app/api/model/contactDto.ts
/**
 * JobSite API.
 * this api has three main resources jobs, contacts and applicants
 *
 * OpenAPI spec version: 0.0.1
 * Contact: seancarlin90@googlemail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ContactDto {
    /**
     * The date time the contact was first created
     */
    createdtime?: Date;
    /**
     * The contact's date of birth
     */
    dob: string;
    /**
     * The contact's first name
     */
    firstname: string;
    /**
     * The contact id (auto generated)
     */
    id?: number;
    /**
     * The contact's last name
     */
    lastname: string;
    /**
     * The date time the contact record was last modified
     */
    modifiedtime?: Date;
    /**
     * The contact's profile picture
     */
    profilePicUrl?: string;
}


export interface ContactListDto {
    contactList: ContactDto[]
}